v1.5.16 TT365 match card type A and Refactor match models and parsing for doubles support
- Update `MatchCard` and `MatchPlayer` to use `double` for set scores.
- Add `Sets` to `MatchCard` and computed properties to `MatchSet`.
- Enhance `Player` with `DisplayName` for substitutes.
- Revise `MatchPage.razor` to use new properties and display logic.
- Refactor `GetMatchCard.cs` to parse sets, players, and Player of the Match.
- Implement `ParseMatchCardTypeA` for extracting set/player data from HTML.

+++ src/Smab.TTInfo.TT365/Models/TT365/MatchPlayer.cs
@@ -8,4 +8,4 @@
 /// <param name="SetsWon">The number of sets won by the player during the match.</param>
 /// <param name="PoM">A value indicating whether the player was awarded Player of the Match. <see langword="true"/> if the player was the
 /// Player of the Match; otherwise, <see langword="false"/>.</param>
-public record MatchPlayer(string Name, int Id, int SetsWon, bool PoM);
+public record MatchPlayer(string Name, int Id, double SetsWon, bool PoM);
